+++ lib/routes.ts
@@ -49,7 +49,7 @@ type RouteHelpers = Collection<RouteHelper>;
 type Configuration = {
   prefix: string;
   default_url_options: RouteParameters;
-  omit_undefined_query_parameters: boolean;
+  omit_undefined_query_parameters: boolean | null;
   special_options_key: string;
   serializer: Serializer;
 };
@@ -81,7 +81,7 @@ declare const RubyVariables: {
   PREFIX: string;
   DEPRECATED_FALSE_PARAMETER_BEHAVIOR: boolean;
   DEPRECATED_NIL_QUERY_PARAMETER_BEHAVIOR: boolean;
-  OMIT_UNDEFINED_QUERY_PARAMETERS: boolean;
+  OMIT_UNDEFINED_QUERY_PARAMETERS: boolean | null;
   SPECIAL_OPTIONS_KEY: string;
   DEFAULT_URL_OPTIONS: RouteParameters;
   SERIALIZER: Serializer;